+++ NetherGate/src/main/java/com/elmakers/mine/bukkit/plugins/nether/dao/Portal.java
@@ -2,10 +2,9 @@ package com.elmakers.mine.bukkit.plugins.nether.dao;
 
 import java.util.Date;
 
-import com.elmakers.mine.bukkit.gameplay.BoundingBox;
 import com.elmakers.mine.bukkit.persistence.annotation.PersistClass;
 import com.elmakers.mine.bukkit.persistence.annotation.PersistField;
-import com.elmakers.mine.bukkit.persistence.dao.WorldData;
+import com.elmakers.mine.bukkit.persistence.dao.LocationData;
 
 @PersistClass(schema="nether", name="portal")
 public class Portal
@@ -42,27 +41,16 @@ public class Portal
 	{
 		this.target = target;
 	}
-	
-	@PersistField(contained=true)
-	public BoundingBox getArea()
-	{
-		return area;
-	}
-	
-	public void setArea(BoundingBox area)
-	{
-		this.area = area;
-	}
 
 	@PersistField
-	public NetherPlayer getOwner()
+	public NetherPlayer getCreator()
 	{
-		return owner;
+		return creator;
 	}
 
-	public void setOwner(NetherPlayer owner)
+	public void setCreator(NetherPlayer creator)
 	{
-		this.owner = owner;
+		this.creator = creator;
 	}
 
 	@PersistField
@@ -110,24 +98,23 @@ public class Portal
 	}
 
 	@PersistField
-	public WorldData getWorld()
+	public LocationData getLocation()
 	{
-		return world;
+		return location;
 	}
 
-	public void setWorld(WorldData world)
+	public void setLocation(LocationData location)
 	{
-		this.world = world;
+		this.location = location;
 	}
 
 	protected int			id;
-	protected WorldData		world;
+	protected LocationData 	location;
 	protected PortalArea	container;
 	protected String		name;
 	protected boolean		active			= false;
 	protected boolean		updatePending	= false;
 	protected Date			lastUsed;
 	protected Portal		target;
-	protected BoundingBox	area;
-	protected NetherPlayer	owner;
+	protected NetherPlayer	creator;
 }